requirements.txt
Flask
mysql-connector-python==8.1.0